requirements.txt
yt-dlp
faster-whisper
demucs
deepgram-sdk
google-generativeai
anthropic
edge-tts
elevenlabs
requests
ffmpeg-python
python-dotenv
tqdm
srt
opencv-python
ollama
Pillow>=10.0

# Optional: higher-quality scene detection for step1b_scenes
# (primary path uses ffmpeg scdet which is always available)
# pip install "scenedetect>=0.6,<1.0"

# Optional: YouTube upload (--upload flag in step7)
# pip install google-api-python-client google-auth-httplib2 google-auth-oauthlib
# google-api-python-client>=2.0.0
# google-auth-httplib2>=0.1.0
# google-auth-oauthlib>=0.5.0
